Move everything about Reversed into their own file
