+++ src/visit/visit.rs
@@ -1,6 +1,9 @@
 //! Graph visitor algorithms.
 //!
 
+mod reversed;
+pub use self::reversed::*;
+
 use fixedbitset::FixedBitSet;
 use std::collections::{
     HashSet,
@@ -51,13 +54,6 @@ pub trait GraphRef : Copy + GraphBase { }
 
 impl<'a, G> GraphRef for &'a G where G: GraphBase { }
 
-impl<G: GraphBase> GraphBase for Reversed<G> {
-    type NodeId = G::NodeId;
-    type EdgeId = G::EdgeId;
-}
-
-impl<G: GraphRef> GraphRef for Reversed<G> { }
-
 impl<'a, G> GraphBase for Frozen<'a, G> where G: GraphBase {
     type NodeId = G::NodeId;
     type EdgeId = G::EdgeId;
@@ -99,10 +95,6 @@ impl<'a, 'b, G> IntoNeighbors for &'b Frozen<'a, G>
 /// Wrapper type for walking the graph as if it is undirected
 #[derive(Copy, Clone)]
 pub struct AsUndirected<G>(pub G);
-
-/// Wrapper type for walking the graph as if all edges are reversed.
-#[derive(Copy, Clone)]
-pub struct Reversed<G>(pub G);
 
 impl<'b, N, E, Ty, Ix> IntoNeighbors for AsUndirected<&'b Graph<N, E, Ty, Ix>> where
     Ty: EdgeType,
@@ -216,19 +208,6 @@ impl<'a, N, E: 'a, Ty, Ix> IntoNodeIdentifiers for &'a StableGraph<N, E, Ty, Ix>
     }
 }
 
-impl<G> IntoNodeIdentifiers for Reversed<G>
-    where G: IntoNodeIdentifiers
-{
-    type NodeIdentifiers = G::NodeIdentifiers;
-    fn node_identifiers(self) -> Self::NodeIdentifiers {
-        self.0.node_identifiers()
-    }
-
-    fn node_count(&self) -> usize {
-        self.0.node_count()
-    }
-}
-
 impl<'a, G> IntoNeighbors for &'a G
     where G: Copy + IntoNeighbors
 {
@@ -249,27 +228,6 @@ impl<'a, G> IntoNeighborsDirected for &'a G
     }
 }
 
-impl<G> IntoNeighbors for Reversed<G>
-    where G: IntoNeighborsDirected
-{
-    type Neighbors = G::NeighborsDirected;
-    fn neighbors(self, n: G::NodeId) -> G::NeighborsDirected
-    {
-        self.0.neighbors_directed(n, Incoming)
-    }
-}
-
-impl<G> IntoNeighborsDirected for Reversed<G>
-    where G: IntoNeighborsDirected
-{
-    type NeighborsDirected = G::NeighborsDirected;
-    fn neighbors_directed(self, n: G::NodeId, d: EdgeDirection)
-        -> G::NeighborsDirected
-    {
-        self.0.neighbors_directed(n, d.opposite())
-    }
-}
-
 /// Access to the graph’s nodes without edges to them (`Incoming`) or from them
 /// (`Outgoing`).
 pub trait IntoExternals : GraphRef {
@@ -286,15 +244,6 @@ impl<'a, N: 'a, E, Ty, Ix> IntoExternals for &'a Graph<N, E, Ty, Ix>
     type Externals = graph::Externals<'a, N, Ty, Ix>;
     fn externals(self, d: EdgeDirection) -> graph::Externals<'a, N, Ty, Ix> {
         Graph::externals(self, d)
-    }
-}
-
-impl<G> IntoExternals for Reversed<G>
-    where G: IntoExternals,
-{
-    type Externals = G::Externals;
-    fn externals(self, d: EdgeDirection) -> G::Externals {
-        self.0.externals(d.opposite())
     }
 }
 
@@ -538,17 +487,6 @@ impl<G: GraphRef> GraphRef for AsUndirected<G> { }
 
 
 impl<G: Visitable> Visitable for AsUndirected<G>
-{
-    type Map = G::Map;
-    fn visit_map(&self) -> G::Map {
-        self.0.visit_map()
-    }
-    fn reset_map(&self, map: &mut Self::Map) {
-        self.0.reset_map(map);
-    }
-}
-
-impl<G: Visitable> Visitable for Reversed<G>
 {
     type Map = G::Map;
     fn visit_map(&self) -> G::Map {
